fix(dashboards): address PR #641 round-4 review feedback
Move builder functions to utils per package-architecture.md,
fix neutral trend icon color mismatch (gray-400 → gray-500).

LFXV2-1644
