+++ apps/lfx-one/src/app/modules/dashboards/marketing-impact/marketing-impact.component.html
@@ -116,7 +116,7 @@
                   } @else if (kpi.trend === 'down') {
                     <i class="fa-solid fa-arrow-down text-xs text-red-600" aria-hidden="true"></i>
                   } @else {
-                    <i class="fa-solid fa-minus text-xs text-gray-400" aria-hidden="true"></i>
+                    <i class="fa-solid fa-minus text-xs text-gray-500" aria-hidden="true"></i>
                   }
                   <span class="text-xs font-medium" [class]="kpi.trendClass">
                     {{ kpi.momChange }}
