+++ packages/shared/src/utils/index.ts
@@ -19,6 +19,7 @@ export * from './platform.utils';
 export * from './project.utils';
 export * from './badge.utils';
 export * from './marketing.utils';
+export * from './marketing-impact.utils';
 export * from './flywheel.utils';
 export * from './rewards.utils';
 export * from './insights.utils';
